requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
black==23.3.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cryptography==40.0.1
daphne==4.0.0
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
msgpack==1.0.5
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
pytest==7.2.2
pytest-asyncio==0.21.0
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.4
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
zope.interface==6.0
